move common components and remove unused page routes
