+++ next-app/app/_components/common/Expandable.tsx
