+++ next-app/app/_components/common/Photo.tsx
