+++ next-app/app/report/_components/ReportForm.tsx
@@ -2,7 +2,7 @@
 
 import { useState, ChangeEvent, FormEvent } from 'react'
 import useSendingStatus from '@/app/_hooks/useSendingStatus'
-import SendButton from '@/components/SendButton'
+import SendButton from '@/app/_components/common/SendButton'
 
 export function ReportForm() {
   const [input, setInput] = useState('')
