+++ next-app/pages/admin/barrios.tsx
@@ -1,8 +1,8 @@
-import Container from '@/components/Container'
-import Layout from '@/components/Layout/Layout'
+import Container from '@/app/_components/common/Container'
+import Layout from '@/app/_components/layout/Layout'
 import { useState, ChangeEvent, FormEvent } from 'react'
 import { useSession } from 'next-auth/react'
-import SendButton from '@/components/SendButton'
+import SendButton from '@/app/_components/common/SendButton'
 import { Barrio } from '.prisma/client'
 import Link from 'next/link'
 import { FaChevronLeft } from 'react-icons/fa'
